Holy crap it works

+++ config/initializers/python_interpreter.rb
@@ -4,24 +4,27 @@ include PythonHelper
 require 'expect'
 require 'pty'
 
-`pkill -f python`
-arr = PTY.spawn("cd py && python")
-$pyout = arr[0]
-$pyin = arr[1]
-$pid = arr[2]
+# Don't run for rake tasks
+unless ENV["RAILS_ENV"].nil? || ENV["RAILS_ENV"] == 'test' || !!@rake
+  `pkill -f python`
+  arr = PTY.spawn("cd py && python")
+  $pyout = arr[0]
+  $pyin = arr[1]
+  $pid = arr[2]
 
-# Wait until terminal is ready
-$pyout.expect(">>>")
+  # Wait until terminal is ready
+  $pyout.expect(">>>")
 
-# Import necessary libs
-init = "
+  # Import necessary libs
+  init = "
 from __future__ import print_function
 import time
 import pychromecast
 import json
-"
+  "
 
-run_py(init)
+  run_py(init)
 
-PythonHelper.refresh_devices()
+  PythonHelper.refresh_devices()
 
+end
